Organise Classes into appropriate packages
