+++ src/main/java/Tyrese.java
@@ -1,9 +1,7 @@
-import exceptions.MarkException;
-import exceptions.TaskException;
 import tasks.*;
-import utils.Greeting;
+import tasks.parser.UserCommandParser;
+import ui.Greeting;
 
-import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.util.Scanner;
