+++ src/main/java/ui/Greeting.java
@@ -1,4 +1,4 @@
-package utils;
+package ui;
 
 public class Greeting {
